+++ src/main/java/com/hsfulda/demo/products/services/InventoryService.java
@@ -1,6 +1,6 @@
 package com.hsfulda.demo.products.services;
 
 public interface InventoryService {
-    int getStockForProductId(int id);
-    boolean reduceStockForProductId(int id, int amount);
+    int getStockForProductId(Long id);
+    boolean reduceStockForProductId(Long id, int amount);
 }
